requirements.txt
defusedxml==0.4.1
Django==1.9.5
gunicorn==19.6.0
oauthlib==2.0.0
PyJWT==1.4.2
python-social-auth==0.2.14
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
whitenoise==3.2.2
dj-database-url==0.4.0
psycopg2==2.6.1
braintree==3.30.0
django-twilio==0.8.0
django-autofixture==0.12.1
django-crispy-forms==1.6.1
django-registration-redux==1.4
django-redis-cache==1.7.1
django-storages-redux==1.3.2
Collectfast==0.5.1
boto==2.45.0
boto3==1.4.4